[window] Implement GetWindowSizeAndRotation and GetWindowSizeAndRotation2
For getting screen size and orientation for given mode

+++ src/emu/epoc/include/epoc/kernel.h
@@ -75,7 +75,7 @@ namespace eka2l1 {
 
     struct find_handle {
         int index;
-        uint64_t object_id;
+        uint32_t object_id;
     };
 
     namespace arm {
@@ -116,12 +116,12 @@ namespace eka2l1 {
 
         uint32_t create_handle_lastest(kernel::owner_type owner);
 
-        mutable std::atomic<uint64_t> uid_counter;
+        mutable std::atomic<uint32_t> uid_counter;
 
         void do_state_of(common::chunkyseri &seri, const kernel::object_type t);
 
     public:
-        uint64_t next_uid() const;
+        uint32_t next_uid() const;
 
         explicit kernel_system()
             : uid_counter(0), libmngr(nullptr),
@@ -249,7 +249,7 @@ namespace eka2l1 {
         int close(uint32_t handle);
 
         kernel_obj_ptr get_kernel_obj(uint32_t handle);
-        kernel_obj_ptr get_kernel_obj_by_id(uint64_t id);
+        kernel_obj_ptr get_kernel_obj_by_id(uint32_t id);
 
         thread_ptr get_thread_by_name(const std::string &name);
         thread_ptr get_thread_by_handle(uint32_t handle);
