+++ src/emu/epoc/src/services/domain/domain.cpp
@@ -312,14 +312,14 @@ namespace eka2l1 {
     }
 
     void domain_server::request_transition_nof(service::ipc_context ctx) {
-        const std::uint64_t sid = ctx.msg->msg_session->unique_id();
+        const std::uint32_t sid = ctx.msg->msg_session->unique_id();
 
         nof_enable[sid] = true;
         ctx.set_request_status(KErrNone);
     }
 
     void domain_server::cancel_transition_nof(service::ipc_context ctx) {
-        const std::uint64_t sid = ctx.msg->msg_session->unique_id();
+        const std::uint32_t sid = ctx.msg->msg_session->unique_id();
 
         nof_enable[sid] = false;
         ctx.set_request_status(KErrNone);
@@ -356,7 +356,7 @@ namespace eka2l1 {
     }
 
     void domain_server::defer_acknowledge(service::ipc_context ctx) {
-        const std::uint64_t ssid = ctx.msg->msg_session->unique_id();
+        const std::uint32_t ssid = ctx.msg->msg_session->unique_id();
         domain_ptr dm = control_domains[ssid];
 
         if (!dm) {
@@ -380,7 +380,7 @@ namespace eka2l1 {
     }
 
     void domain_server::cancel_defer_acknowledge(service::ipc_context ctx) {
-        const std::uint64_t ssid = ctx.msg->msg_session->unique_id();
+        const std::uint32_t ssid = ctx.msg->msg_session->unique_id();
         domain_ptr dm = control_domains[ssid];
 
         if (!dm) {
